server/requirements/base.txt
django>=1.8,<1.9
#django-environ==0.4.0
#django-flat-theme==1.1.3

#pillow==3.0.0
#easy-thumbnails==2.2.1
#django-image-cropping==1.0.2

#django-reversion>=1.8,<1.9

psycopg2==2.6.1

#django-extensions==1.5.5

#django-ckeditor==5.0.3

#django-compressor==1.5

#django-tagulous==0.11.1

#django-localflavor==1.1

#django-colorful==1.1.0

#django-geoposition==0.2.2

requests==2.9.1

djangorestframework==3.3.2
djangorestframework-recursive==0.1.1

#wagtail==1.3.1

#modules required for the jobvite synchronization script
#bleach==1.4.1

#for Instagram Integration
#httplib2==0.9.2

#django-versatileimagefield==1.3

# USER AGENT detection
#PyYAML==3.11
#ua-parser==0.6.1
user-agents==1.0.1
django-user-agents==0.3.0

python-memcached==1.57

# belongs in develop but breaks ci if missing
ipdb==0.9.0

#JIRA API
jira

#pycrypto
pycrypto